Unit test credit modifier

+++ src/main/java/com/synacy/trainee/leavemanagementsystem/leaveapplication/LeaveApplication.java
@@ -34,7 +34,7 @@ public class LeaveApplication {
     @CreationTimestamp
     private LocalDate appliedDate;
     @Column
-    private double numberOfDays;
+    private Integer numberOfDays;
 
     @Column
     private String reason;
